Found WPI Arcade Controller
I found an ArcadeController built into the WPILib stuff.  I have
implemented that in a rough layout.

Also cleaned up a bunch of other stuff.  Moved object instantiation of
the controllers and the drivetrain as well as intake to the “init”
method of the robot.  Cameras and stuff can go there as well later.

Also cleaned up the InTake class to what we have/need for now.